Adjustments in jupyter
created a separated notebook to avoid the arima experiemntations once it was pointed that it was not the best model (must apply sarima first to consider the seasonality present in this dataset). this is also a way to reset, refresh and reorganise the notebook to have the sections displayed more clearly and facilitate my understandinf of what is happening and has to still be done.